Cleanup after viewmodel refactoring
