+++ WoWsShipBuilder.UI/UserControls/ShipModuleControl.axaml
@@ -5,11 +5,12 @@
              xmlns:converters="clr-namespace:WoWsShipBuilder.UI.Converters"
              xmlns:data="clr-namespace:WoWsShipBuilder.DataStructures;assembly=WoWsShipBuilder.DataStructures"
              xmlns:viewModels="clr-namespace:WoWsShipBuilder.UI.ViewModels;assembly=WoWsShipBuilder.ViewModels"
+             xmlns:shipVm="clr-namespace:WoWsShipBuilder.ViewModels.ShipVm;assembly=WoWsShipBuilder.ViewModels"
              mc:Ignorable="d" d:DesignWidth="800" d:DesignHeight="450"
              x:Class="WoWsShipBuilder.UI.UserControls.ShipModuleControl"
              Name="ShipModuleControl">
     <Design.DataContext>
-        <viewModels:ShipModuleViewModel />
+        <shipVm:ShipModuleViewModel />
     </Design.DataContext>
     <UserControl.Resources>
         <converters:IsModuleSelectedConverter x:Key="ModuleSelectedConverter" />
