+++ WoWsShipBuilder.UI/Views/MainWindow.axaml
@@ -6,11 +6,9 @@
         x:Class="WoWsShipBuilder.UI.Views.MainWindow"
         xmlns:translations="clr-namespace:WoWsShipBuilder.UI.Translations;assembly=WoWsShipBuilder.Core"
         xmlns:userControls="clr-namespace:WoWsShipBuilder.UI.UserControls"
-        xmlns:vm="clr-namespace:WoWsShipBuilder.UI.ViewModels"
         xmlns:conv="clr-namespace:WoWsShipBuilder.UI.Converters"
         xmlns:utilities="clr-namespace:WoWsShipBuilder.UI.Utilities"
         xmlns:shipVm="clr-namespace:WoWsShipBuilder.UI.ViewModels.ShipVm"
-        xmlns:vm1="clr-namespace:WoWsShipBuilder.ViewModels.ShipVm;assembly=WoWsShipBuilder.ViewModels"
         Title="WoWs Ship Builder"
         Height="850" Width="1775" CanResize="True"
         MinHeight="850" MinWidth="1775"
@@ -24,7 +22,6 @@
     <Window.Resources>
         <conv:ImagePathConverter x:Key="ImagePathConverter" />
         <conv:ClassToPathConverter x:Key="PathConverter" />
-        <conv:EnumConverter x:Key="EnumConverter" />
         <conv:DecimalToRomanConverter x:Key="RomanConverter" />
     </Window.Resources>
 
@@ -108,7 +105,7 @@
                 </Border>
 
                 <!--SecondRow-->
-                <Grid Grid.Row="1" ColumnDefinitions="3*,Auto,4.5*,*" Margin="0,0,0,10">
+                <Grid Grid.Row="1" ColumnDefinitions="3*,Auto,5.5*" Margin="0,0,0,10">
                     <!--Captain and flag selector-->
                     <Grid RowDefinitions="Auto,*" Grid.Column="0" Margin="10,0,0,0" VerticalAlignment="Stretch">
                         <Border Classes="BorderPanel" Grid.Row="0">
@@ -139,53 +136,6 @@
                     <Border Classes="BorderPanel" Grid.Column="2" Padding="5" Margin="0,0,10,0" CornerRadius="5">
                         <userControls:ShipStatsControl DataContext="{Binding ShipStatsControlViewModel}" />
                     </Border>
-                    <!--XP calculator-->
-                    <Grid Grid.Column="3" RowDefinitions="*, Auto" Margin="0,0,10,0">
-                        <Border Classes="BorderPanel" Grid.Row="0">
-                            <StackPanel Orientation="Vertical" Margin="5">
-                                <TextBlock Text="{x:Static translations:Translation.MainWindow_XpCalculator}" Margin="0,0,0,10" />
-                                <TextBlock Text="{x:Static translations:Translation.MainWindow_BaseXp}" />
-                                <TextBox Text="{Binding BaseXp}" MaxLength="5" Name="BaseXpInput" Margin="0,0,0,10" />
-                                <TextBlock Text="{x:Static translations:Translation.MainWindow_XpBonus}" />
-                                <TextBox Text="{Binding XpBonus}" MaxLength="4" Name="XpInput" Margin="0,0,0,10" />
-                                <TextBlock Text="{x:Static translations:Translation.MainWindow_CommanderXpBonus}" />
-                                <TextBox Text="{Binding CommanderXpBonus}" MaxLength="4" Name="CommanderXpInput" Margin="0,0,0,10" />
-                                <TextBlock Text="{x:Static translations:Translation.MainWindow_FreeXpBonus}" />
-                                <TextBox Text="{Binding FreeXpBonus}" MaxLength="4" Name="FreeXpInput" Margin="0,0,0,10" />
-                                <RadioButton GroupName="AccountType"
-                                             IsChecked="{Binding AccountType, Converter={StaticResource EnumConverter}, ConverterParameter={x:Static vm1:Account.Normal}}"
-                                             Margin="0,0,0,5">
-                                    <TextBlock Text="{x:Static translations:Translation.MainWindow_NormalAccount}" TextWrapping="Wrap" />
-                                </RadioButton>
-                                <RadioButton GroupName="AccountType"
-                                             IsChecked="{Binding AccountType, Converter={StaticResource EnumConverter}, ConverterParameter={x:Static vm1:Account.WGPremium}}"
-                                             Margin="0,0,0,5">
-                                    <TextBlock Text="{x:Static translations:Translation.MainWindow_WGPremiumAccount}" TextWrapping="Wrap" />
-                                </RadioButton>
-                                <RadioButton GroupName="AccountType"
-                                             IsChecked="{Binding AccountType, Converter={StaticResource EnumConverter}, ConverterParameter={x:Static vm1:Account.WoWsPremium}}"
-                                             Margin="0,0,0,5">
-                                    <TextBlock Text="{x:Static translations:Translation.MainWindow_WoWsPremiumAccount}" TextWrapping="Wrap" />
-                                </RadioButton>
-                                <Separator Margin="0,0,0,5" />
-                                <StackPanel Orientation="Horizontal" Margin="0,0,0,10">
-                                    <TextBlock Text="{x:Static translations:Translation.MainWindow_Xp}" />
-                                    <TextBlock Text="{Binding Xp}" TextWrapping="Wrap" Margin="5,0,0,0" />
-                                </StackPanel>
-                                <StackPanel Orientation="Horizontal" Margin="0,0,0,10">
-                                    <TextBlock Text="{x:Static translations:Translation.MainWindow_CommanderXp}" />
-                                    <TextBlock Text="{Binding CommanderXp}" TextWrapping="Wrap" Margin="5,0,0,0" />
-                                </StackPanel>
-                                <StackPanel Orientation="Horizontal">
-                                    <TextBlock Text="{x:Static translations:Translation.MainWindow_FreeXp}" />
-                                    <TextBlock Text="{Binding FreeXp}" TextWrapping="Wrap" Margin="5,0,0,0" />
-                                </StackPanel>
-                            </StackPanel>
-                        </Border>
-                        <Border Classes="BorderPanel" Grid.Row="1" Margin="0,10,0,0" Padding="10">
-                            <Image Source="/Assets/Art/logo_white_text.png" />
-                        </Border>
-                    </Grid>
                 </Grid>
             </Grid>
         </Grid>
