+++ WoWsShipBuilder.UI/Views/MainWindow.axaml.cs
@@ -2,20 +2,17 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Reactive;
-using System.Text.RegularExpressions;
 using Avalonia;
 using Avalonia.Controls;
 using Avalonia.Controls.ApplicationLifetimes;
 using Avalonia.Controls.Mixins;
 using Avalonia.Input;
-using Avalonia.Interactivity;
 using Avalonia.Markup.Xaml;
 using ReactiveUI;
 using WoWsShipBuilder.DataStructures;
 using WoWsShipBuilder.UI.Translations;
 using WoWsShipBuilder.UI.UserControls;
 using WoWsShipBuilder.UI.Utilities;
-using WoWsShipBuilder.UI.ViewModels;
 using WoWsShipBuilder.UI.ViewModels.ShipVm;
 using WoWsShipBuilder.ViewModels.Helper;
 
@@ -23,8 +20,6 @@ namespace WoWsShipBuilder.UI.Views
 {
     public partial class MainWindow : ScalableReactiveWindow<MainWindowViewModel>
     {
-        private static readonly Regex Regex = new("\\D+");
-
         public MainWindow()
         {
             InitializeComponent();
@@ -90,29 +85,12 @@ namespace WoWsShipBuilder.UI.Views
         private void InitializeComponent()
         {
             AvaloniaXamlLoader.Load(this);
-            var baseXp = this.FindControl<TextBox>("BaseXpInput");
-            baseXp.AddHandler(TextInputEvent, AcceptOnlyNumber, RoutingStrategies.Tunnel);
-            var xpInput = this.FindControl<TextBox>("XpInput");
-            xpInput.AddHandler(TextInputEvent, AcceptOnlyNumber, RoutingStrategies.Tunnel);
-            var commanderXpInput = this.FindControl<TextBox>("CommanderXpInput");
-            commanderXpInput.AddHandler(TextInputEvent, AcceptOnlyNumber, RoutingStrategies.Tunnel);
-            var freeXpInput = this.FindControl<TextBox>("FreeXpInput");
-            freeXpInput.AddHandler(TextInputEvent, AcceptOnlyNumber, RoutingStrategies.Tunnel);
         }
 
         protected override void OnOpened(EventArgs e)
         {
             base.OnOpened(e);
             this.HandleAndCheckScaling();
-        }
-
-        public void AcceptOnlyNumber(object? sender, TextInputEventArgs e)
-        {
-            var text = e.Text!;
-            if (Regex.IsMatch(text))
-            {
-                e.Handled = true;
-            }
         }
 
         public void OnClickChangeShipNext(object sender, PointerReleasedEventArgs e)
