+++ WoWsShipBuilder.UI/Views/ShipSelectionWindow.axaml
@@ -5,12 +5,13 @@
         xmlns:translations="clr-namespace:WoWsShipBuilder.UI.Translations;assembly=WoWsShipBuilder.Core"
         xmlns:userControls="clr-namespace:WoWsShipBuilder.UI.UserControls"
         xmlns:viewModels="clr-namespace:WoWsShipBuilder.UI.ViewModels;assembly=WoWsShipBuilder.ViewModels"
+        xmlns:helper="clr-namespace:WoWsShipBuilder.ViewModels.Helper;assembly=WoWsShipBuilder.ViewModels"
         mc:Ignorable="d" d:DesignWidth="500" d:DesignHeight="450"
         x:Class="WoWsShipBuilder.UI.Views.ShipSelectionWindow" Height="450" Width="500"
         GotFocus="InputElement_OnGotFocus"
         WindowStartupLocation="CenterOwner" CanResize="False" ShowInTaskbar="False" SizeToContent="Height">
     <Design.DataContext>
-        <viewModels:ShipSelectionWindowViewModel />
+        <helper:ShipSelectionWindowViewModel />
     </Design.DataContext>
     
     <Window.Styles>
