+++ WoWsShipBuilder.UI/Views/ShipSelectionWindow.axaml.cs
@@ -8,6 +8,7 @@ using Avalonia.Markup.Xaml;
 using Avalonia.ReactiveUI;
 using ReactiveUI;
 using WoWsShipBuilder.UI.ViewModels;
+using WoWsShipBuilder.ViewModels.Helper;
 
 namespace WoWsShipBuilder.UI.Views
 {
