+++ WoWsShipBuilder.UI/Views/SplashScreen.axaml
@@ -6,13 +6,14 @@
         x:Class="WoWsShipBuilder.UI.Views.SplashScreen"
         xmlns:translations="clr-namespace:WoWsShipBuilder.UI.Translations;assembly=WoWsShipBuilder.Core"
         xmlns:viewModels="clr-namespace:WoWsShipBuilder.UI.ViewModels;assembly=WoWsShipBuilder.ViewModels"
+        xmlns:other="clr-namespace:WoWsShipBuilder.ViewModels.Other;assembly=WoWsShipBuilder.ViewModels"
         Title="SplashScreenWindow"
         Name="SplashScreenWindow"
         Height="566" Width="800" CanResize="False"
         SystemDecorations="None"
         WindowStartupLocation="CenterScreen">
     <Design.DataContext>
-        <viewModels:SplashScreenViewModel />
+        <other:SplashScreenViewModel />
     </Design.DataContext>
 
     <RelativePanel Width="800">
