+++ WoWsShipBuilder.UI/Views/SplashScreen.axaml.cs
@@ -10,6 +10,7 @@ using WoWsShipBuilder.Core.Extensions;
 using WoWsShipBuilder.Core.Services;
 using WoWsShipBuilder.UI.Extensions;
 using WoWsShipBuilder.UI.ViewModels;
+using WoWsShipBuilder.ViewModels.Other;
 
 namespace WoWsShipBuilder.UI.Views
 {
