+++ WoWsShipBuilder.UI/Views/ValueSelectionWindow.axaml.cs
@@ -4,6 +4,7 @@ using Avalonia.Markup.Xaml;
 using Avalonia.ReactiveUI;
 using ReactiveUI;
 using WoWsShipBuilder.UI.ViewModels;
+using WoWsShipBuilder.ViewModels.Helper;
 
 namespace WoWsShipBuilder.UI.Views
 {
