+++ WoWsShipBuilder.ViewModels/Other/SettingsWindowViewModelBase.cs
