+++ WoWsShipBuilder.ViewModels/ShipVm/FiringAngleViewModelBase.cs
