+++ WoWsShipBuilder.ViewModels/ShipVm/ShipModuleViewModel.cs
@@ -8,7 +8,7 @@ using WoWsShipBuilder.Core.DataProvider;
 using WoWsShipBuilder.DataStructures;
 using WoWsShipBuilder.ViewModels.Base;
 
-namespace WoWsShipBuilder.UI.ViewModels
+namespace WoWsShipBuilder.ViewModels.ShipVm
 {
     public class ShipModuleViewModel : ViewModelBase, IBuildStorable
     {
