+++ WoWsShipBuilder.ViewModels/ShipVm/ShipStatsControlViewModelBase.cs
