+++ WoWsShipBuilder.ViewModels/ShipVm/SignalSelectorViewModel.cs
@@ -9,7 +9,7 @@ using WoWsShipBuilder.Core.DataProvider;
 using WoWsShipBuilder.DataStructures;
 using WoWsShipBuilder.ViewModels.Base;
 
-namespace WoWsShipBuilder.UI.ViewModels
+namespace WoWsShipBuilder.ViewModels.ShipVm
 {
     public class SignalSelectorViewModel : ViewModelBase
     {
