+++ WoWsShipBuilder.ViewModels/ShipVm/SkillActivationItemViewModel.cs
@@ -2,7 +2,7 @@ using System.Collections.Generic;
 using ReactiveUI;
 using WoWsShipBuilder.ViewModels.Base;
 
-namespace WoWsShipBuilder.UI.ViewModels
+namespace WoWsShipBuilder.ViewModels.ShipVm
 {
     public class SkillActivationItemViewModel : ViewModelBase
     {
